Extract HTTP related classes to their own namespace
