+++ src/ClientFactory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Werkspot\KvkApi;
 
-use Werkspot\KvkApi\Client\Adapter\AdapterInterface;
+use Werkspot\KvkApi\Http\ClientInterface;
 use Werkspot\KvkApi\Client\Builder\Profile\Company\AddressBuilder;
 use Werkspot\KvkApi\Client\Builder\Profile\Company\BusinessActivityBuilder;
 use Werkspot\KvkApi\Client\Builder\Profile\Company\TradeNamesBuilder;
@@ -13,7 +13,7 @@ use Werkspot\KvkApi\Client\Builder\ProfileResponseBuilder;
 
 final class ClientFactory
 {
-    public static function getClient(AdapterInterface $adapter): Client
+    public static function getClient(ClientInterface $adapter): Client
     {
         $profileResponseBuilder =  new ProfileResponseBuilder(
             new CompanyBuilder(
