+++ src/Http/ClientInterface.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Werkspot\KvkApi\Client\Adapter;
+namespace Werkspot\KvkApi\Http;
 
 use Psr\Http\Message\ResponseInterface;
 use Werkspot\KvkApi\Client\Search\QueryInterface;
 
-interface AdapterInterface
+interface ClientInterface
 {
     public function getEndpoint(string $endpoint, QueryInterface $searchQuery): ResponseInterface;
     public function getUrl(string $url): ResponseInterface;
